% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosSim.R
\name{CosSim}
\alias{CosSim}
\title{1. Cosine Similarity.}
\usage{
CosSim(matrix_cs)
}
\arguments{
\item{matrix_cs}{only required parameter. If required similarity by columns, transpose the input matrix.}
}
\description{
This function calculates the row similarity of a nxm matrix using the cosine pairwise projection
}
\examples{
matrix(data = c(1,1,1,0,1,1,0,1), ncol = 2)
}
\keyword{netowrks}
\keyword{recommendation}
\keyword{similarity}
